+++ packages/toolkit/src/lib/use-instill-form/components/component-output/AudiosField.tsx
@@ -1,40 +1,17 @@
-import { getAudioTypeFromBase64 } from "../../../../view";
 import { Nullable } from "../../../type";
-import { ConnectorNodeFieldRoot, EndNodeFieldRoot } from "./FieldRoot";
+import { FieldRoot } from "./FieldRoot";
 
 export type AudiosFieldProps = {
-  nodeType: "end" | "connector";
   title: Nullable<string>;
   audios: Nullable<string[]>;
   hideField?: boolean;
 };
 
 export const AudiosField = (props: AudiosFieldProps) => {
-  const { nodeType, title, audios, hideField } = props;
-
-  if (nodeType === "connector") {
-    return (
-      <ConnectorNodeFieldRoot title={title} fieldKey={`${title}-field`}>
-        {audios && !hideField ? (
-          <div className="flex w-full flex-col">
-            {audios.map((audio) => {
-              return (
-                <audio
-                  key={`${title}-${audio}-field`}
-                  className="w-full"
-                  controls={true}
-                  src={audio}
-                />
-              );
-            })}
-          </div>
-        ) : null}
-      </ConnectorNodeFieldRoot>
-    );
-  }
+  const { title, audios, hideField } = props;
 
   return (
-    <EndNodeFieldRoot title={title} key={`${title}-field`}>
+    <FieldRoot title={title} fieldKey={`${title}-field`}>
       {audios && !hideField ? (
         <div className="flex w-full flex-col">
           {audios.map((audio) => {
@@ -49,6 +26,6 @@ export const AudiosField = (props: AudiosFieldProps) => {
           })}
         </div>
       ) : null}
-    </EndNodeFieldRoot>
+    </FieldRoot>
   );
 };