+++ packages/toolkit/src/lib/use-instill-form/components/component-output/FieldRoot.tsx
@@ -2,7 +2,7 @@ import cn from "clsx";
 import * as React from "react";
 import { Nullable } from "../../../type";
 
-export const ConnectorNodeFieldRoot = ({
+export const FieldRoot = ({
   fieldKey,
   title,
   children,
@@ -17,38 +17,11 @@ export const ConnectorNodeFieldRoot = ({
     <div
       key={fieldKey}
       className={cn(
-        "flex w-full max-w-[calc(var(--pipeline-builder-node-available-width)-var(--pipeline-builder-node-padding-x))] flex-row flex-wrap gap-y-2 rounded-[6px] bg-semantic-bg-primary p-2",
-        className
-      )}
-    >
-      <p className="text-semantic-fg-secondary product-body-text-4-semibold">
-        {title}
-      </p>
-      {children}
-    </div>
-  );
-};
-
-export const EndNodeFieldRoot = ({
-  key,
-  title,
-  children,
-  className,
-}: {
-  key: string;
-  title: Nullable<string>;
-  children: React.ReactNode;
-  className?: string;
-}) => {
-  return (
-    <div
-      key={key}
-      className={cn(
         "flex w-full flex-col gap-y-2 rounded-[6px] bg-semantic-bg-primary p-2",
         className
       )}
     >
-      <p className="text-semantic-fg-primary product-body-text-3-semibold">
+      <p className="text-semantic-fg-primary product-body-text-4-semibold">
         {title}
       </p>
       {children}