+++ packages/toolkit/src/lib/use-instill-form/components/component-output/ImagesField.tsx
@@ -1,41 +1,17 @@
 import { Nullable } from "../../../type";
-import { ConnectorNodeFieldRoot, EndNodeFieldRoot } from "./FieldRoot";
+import { FieldRoot } from "./FieldRoot";
 
 export type ImagesFieldProps = {
-  nodeType: "end" | "connector";
   title: Nullable<string>;
   images: Nullable<string>[];
   hideField?: boolean;
 };
 
 export const ImagesField = (props: ImagesFieldProps) => {
-  const { nodeType, title, images, hideField } = props;
-
-  if (nodeType === "connector") {
-    return (
-      <ConnectorNodeFieldRoot title={title} fieldKey={`${title}-field`}>
-        {images && !hideField ? (
-          <div className="flex w-full flex-wrap">
-            {images?.slice(0, 5)?.map((image) => {
-              if (!image) return null;
-
-              return (
-                <img
-                  key={`${title}-${image}-field`}
-                  alt={`${title}-images-{idx}`}
-                  src={image}
-                  className="object-contain"
-                />
-              );
-            })}
-          </div>
-        ) : null}
-      </ConnectorNodeFieldRoot>
-    );
-  }
+  const { title, images, hideField } = props;
 
   return (
-    <EndNodeFieldRoot title={title} key={`${title}-field`}>
+    <FieldRoot title={title} fieldKey={`${title}-field`}>
       {images && !hideField ? (
         <div className="flex w-full flex-wrap">
           {images?.slice(0, 5)?.map((image) => {
@@ -52,6 +28,6 @@ export const ImagesField = (props: ImagesFieldProps) => {
           })}
         </div>
       ) : null}
-    </EndNodeFieldRoot>
+    </FieldRoot>
   );
 };