+++ packages/toolkit/src/lib/use-instill-form/components/component-output/ObjectsField.tsx
@@ -1,37 +1,17 @@
 import { GeneralRecord, Nullable } from "../../../type";
-import { ConnectorNodeFieldRoot, EndNodeFieldRoot } from "./FieldRoot";
+import { FieldRoot } from "./FieldRoot";
 
 export type ObjectsFieldProps = {
-  nodeType: "end" | "connector";
   title: Nullable<string>;
   objects: Nullable<GeneralRecord[]>;
   hideField?: boolean;
 };
 
 export const ObjectsField = (props: ObjectsFieldProps) => {
-  const { nodeType, title, objects, hideField } = props;
-
-  if (nodeType === "connector") {
-    return (
-      <ConnectorNodeFieldRoot title={title} fieldKey={`${title}-field`}>
-        {objects && !hideField ? (
-          <div className="flex w-full flex-col flex-wrap gap-2">
-            {objects?.map((object) => (
-              <div
-                key={`${title}-${JSON.stringify(object)}-field`}
-                className="flex min-h-[36px] w-full whitespace-pre-wrap break-all rounded border border-semantic-bg-line p-1 text-semantic-fg-primary product-body-text-4-regular"
-              >
-                {JSON.stringify(object, null, 4)}
-              </div>
-            ))}
-          </div>
-        ) : null}
-      </ConnectorNodeFieldRoot>
-    );
-  }
+  const { title, objects, hideField } = props;
 
   return (
-    <EndNodeFieldRoot title={title} key={`${title}-field`}>
+    <FieldRoot title={title} fieldKey={`${title}-field`}>
       {objects && !hideField ? (
         <div className="flex w-full flex-col flex-wrap gap-2">
           {objects?.map((object) => (
@@ -44,6 +24,6 @@ export const ObjectsField = (props: ObjectsFieldProps) => {
           ))}
         </div>
       ) : null}
-    </EndNodeFieldRoot>
+    </FieldRoot>
   );
 };