+++ packages/toolkit/src/lib/use-instill-form/pick/pickRegularFieldsFromInstillFormTree.tsx
@@ -187,6 +187,7 @@ export function pickRegularFieldsFromInstillFormTree(
     if (enableSmartHint) {
       return (
         <SmartHintFields.TextArea
+          fieldKey={tree.fieldKey}
           key={tree.path}
           path={tree.path}
           form={form}
@@ -238,7 +239,7 @@ export function pickRegularFieldsFromInstillFormTree(
   if (enableSmartHint) {
     return (
       <SmartHintFields.TextField
-        key={tree.path}
+        fieldKey={tree.fieldKey}
         path={tree.path}
         form={form}
         title={title}