+++ packages/toolkit/src/view/pipeline-builder/components/NodeWrapper.tsx
@@ -8,6 +8,7 @@ const selector = (store: InstillStore) => ({
   updateNodes: store.updateNodes,
   updatePipelineRecipeIsDirty: store.updatePipelineRecipeIsDirty,
   selectedConnectorNodeId: store.selectedConnectorNodeId,
+  updateSelectedConnectorNodeId: store.updateSelectedConnectorNodeId,
 });
 
 export const NodeWrapper = ({
@@ -27,18 +28,27 @@ export const NodeWrapper = ({
   renderBottomBarInformation?: () => React.ReactNode;
   renderNodeBottomBar?: () => React.ReactNode;
 }) => {
-  const { updateNodes, updatePipelineRecipeIsDirty, selectedConnectorNodeId } =
-    useInstillStore(useShallow(selector));
+  const {
+    updateNodes,
+    updatePipelineRecipeIsDirty,
+    selectedConnectorNodeId,
+    updateSelectedConnectorNodeId,
+  } = useInstillStore(useShallow(selector));
   const timer = React.useRef<Nullable<number>>(null);
   const [noteValue, setNoteValue] = React.useState(note);
 
   return (
-    <div className="relative">
+    <div
+      className="relative"
+      onClick={() => {
+        updateSelectedConnectorNodeId(() => id);
+      }}
+    >
       <div
         className={cn(
           "absolute left-0 top-0 w-[var(--pipeline-builder-node-available-width)] rounded border border-semantic-warning-default bg-semantic-warning-bg p-2",
           noteIsOpen ? "" : "hidden",
-          "-translate-y-[calc(100%+25px)]"
+          "-translate-y-[calc(100%+16px)]"
         )}
       >
         <Textarea
@@ -86,7 +96,7 @@ export const NodeWrapper = ({
       {renderBottomBarInformation ? (
         <div
           id={`${id}-node-bottom-information-container`}
-          className="nodrag nowheel absolute bottom-0 left-0 w-[var(--pipeline-builder-node-available-width)] translate-y-[calc(100%+25px)]"
+          className="nodrag nowheel absolute bottom-0 left-0 w-[var(--pipeline-builder-node-available-width)] translate-y-[calc(100%+16px)]"
         >
           {renderBottomBarInformation()}
         </div>